Move ILM functions out of common.rb
Move into their own module to be mixed in to elasticsearch.rb
